test: use client interface in redpanda service
This allows the redpanda service to use any kafka client implementation
for basic admin tasks.
