+++ tests/rptest/tests/librdkafka_test.py
@@ -38,7 +38,8 @@ class LibrdkafkaTest(Test):
 
     def _start_redpanda(self, num_brokers):
         self._redpanda = RedpandaService(self._context,
-                                         num_brokers=num_brokers,
+                                         num_brokers,
+                                         KafkaCliTools,
                                          extra_rp_conf=self._extra_rp_conf)
         self._redpanda.start()
